+++ src/xbrl_us/app.py
@@ -1,16 +1,25 @@
+from pathlib import Path
+
 import streamlit as st
 
 from xbrl_us import XBRL
 
+user_info_path = Path.home() / ".xbrl-us"
 
-def try_credentials(user_name: str, pass_word: str, client_id: str, client_secret: str):
+
+def try_credentials(user_name: str, pass_word: str, client_id: str, client_secret: str, store: bool = False):
     try:
+        if store:
+            store = "y"
+        else:
+            store = "n"
         with st.spinner(text="Validating credentials..."):
-            XBRL(username=user_name, password=pass_word, client_id=client_id, client_secret=client_secret)._get_token()
+            XBRL(username=user_name, password=pass_word, client_id=client_id, client_secret=client_secret)._get_token(store=store)
             st.session_state.username = user_name
             st.session_state.password = pass_word
             st.session_state.client_id = client_id
             st.session_state.client_secret = client_secret
+            st.session_state.returning_user = True
     except Exception as e:
         st.error(f"Invalid credentials. Please try again. {e}")
         st.stop()
@@ -58,6 +67,13 @@ def show_login():
         help="Your client secret for the [XBRL.US](https://www.xbrl.us) API.",
     )
 
+    # checkbox for remember me
+    remember_me = st.checkbox(
+        label="Remember me",
+        value=False,
+        key="remember_me",
+    )
+
     disable_login_btn = False
     if username == "" or password == "" or client_id == "" or client_secret == "":
         disable_login_btn = True
@@ -68,9 +84,10 @@ def show_login():
         use_container_width=True,
         disabled=disable_login_btn,
     )
+
     if verify_api:
         # try the credentials before creating xbrl object
-        try_credentials(user_name=username, pass_word=password, client_id=client_id, client_secret=client_secret)
+        try_credentials(user_name=username, pass_word=password, client_id=client_id, client_secret=client_secret, store=remember_me)
         st.experimental_rerun()
 
 
@@ -110,8 +127,6 @@ def range_and_slider_for_array_integers(key, values):
         label_visibility="collapsed",
         on_change=lambda: st.session_state.pop(f"{key}"),
     )
-    if f"{key}_max_value" not in st.session_state:
-        st.session_state[f"{key}_max_value"] = values["max"]
 
     if st.session_state[f"{key}_input_method"] == "Range":
         # update_slider_range("period.fiscal-year_input")
@@ -134,7 +149,7 @@ def range_and_slider_for_array_integers(key, values):
                 key=f"{key}_max_value",
             )
 
-            st.session_state[key] = range(st.session_state[f"{key}_min_value"], st.session_state[f"{key}_max_value"])
+            st.session_state[key] = range(st.session_state[f"{key}_min_value"], st.session_state[f"{key}_max_value"] + 1)
 
         else:
             st.slider(
@@ -160,11 +175,17 @@ def range_and_slider_for_array_integers(key, values):
         )
 
 
-def text_box_for_array_strings_no_ops(key):
+def text_box_for_array_strings_no_ops(key, palceholder):
     st.text_area(
         label=f"**{key}**",
+        placeholder=f"{palceholder}",
         key=f"{key}",
     )
+
+
+def restart_everything():
+    st.session_state.clear()
+    st.session_state.update({"returning_user": False})
 
 
 if __name__ == "__main__":
@@ -178,19 +199,30 @@ if __name__ == "__main__":
     st.title("Explore XBRL.us Data")
 
     sidebar = st.sidebar
-    if "username" not in st.session_state:
+    if user_info_path.exists():
+        if "returning_user" not in st.session_state:
+            st.session_state.returning_user = True
+            temp_xbrl = XBRL()
+            # show button to continue with the username
+            st.session_state.username = temp_xbrl.username
+            st.session_state.password = temp_xbrl.password
+            st.session_state.client_id = temp_xbrl.client_id
+            st.session_state.client_secret = temp_xbrl.client_secret
+
+    if "returning_user" not in st.session_state or not st.session_state.returning_user:
         st.error("Please enter your credentials to begin.")
 
         with sidebar:
             show_login()
         st.stop()
-    else:
+    if "returning_user" in st.session_state and st.session_state.returning_user:
         with sidebar:
+            st.success(f"Logged in as {st.session_state.username}")
             st.button(
                 label="Log out",
                 type="secondary",
                 use_container_width=True,
-                on_click=lambda: st.session_state.clear(),
+                on_click=lambda: restart_everything(),
                 key="logout",
             )
 
@@ -200,15 +232,21 @@ if __name__ == "__main__":
             client_id=st.session_state.client_id,
             client_secret=st.session_state.client_secret,
         )
+        if "account_limit" in st.session_state:
+            xbrl.account_limit = st.session_state.account_limit
 
         st.session_state.methods = xbrl.methods()
 
         method = sidebar.selectbox(
             label="API Method",
-            options=sorted(st.session_state.methods),
-            index=19,
+            options=[
+                "fact search",
+                "report search",
+                "entity search",
+            ],  # sorted(st.session_state.methods), TODO: This is hard coded. Need to fix it.
+            index=0,
             key="method",
-            disabled=True,
+            disabled=False,
             help="""Select the method you would like to use.
             For more information on the methods,
             see the [XBRL.US API Documentation](https://xbrlus.github.io/xbrl-api/#/).""",
@@ -219,22 +257,40 @@ if __name__ == "__main__":
         # parameters_options = dict(sorted(method_params.parameters.items(), key=lambda x: x[1]['type']))
         # print the name of the method
         st.header(method)
-        st.markdown(st.session_state.method_params.description)
-        st.caption(f"**API end-point**: {xbrl.acceptable_params(method).url}")
-        # print the url of the method
+
+        # two tabs
+        method_summary, definitions = st.tabs(["Summary", "Field Definition"])
+
+        with method_summary:
+            st.markdown(st.session_state.method_params.description)
+            st.markdown(
+                f"{st.session_state.method_params.long_description} <sup>[1]({st.session_state.method_params.reference})</sup>",
+                unsafe_allow_html=True,
+            )
+            st.markdown(f"**Glossary**: {st.session_state.method_params.reference}")
+            st.caption(f"**API end-point**: {xbrl.acceptable_params(method).url}")
+            # print the url of the method
+
+        with definitions:
+            with st.expander("Search Definitions", expanded=True):
+                if "fields" in st.session_state and len(st.session_state.fields) > 0:
+                    st.markdown(f"**Field Name**: *{st.session_state.fields[-1]}*")
+                    st.markdown(f"**Description**: {xbrl.define(st.session_state.fields[-1])['description']}")
+                else:
+                    st.info("Select a field to see the definition")
 
         # show the list of fields in the sidebar
         with st.container():
             sidebar.multiselect(
-                label="Fields :red[*]",
-                options=st.session_state.method_params.fields,
-                key="fields",
-            )
-
-            sidebar.multiselect(
                 label="Parameters",
                 options=st.session_state.method_params.parameters,
                 key="parameters",
+            )
+
+            sidebar.multiselect(
+                label="Fields :red[*]",
+                options=st.session_state.method_params.fields,
+                key="fields",
             )
 
             sidebar.multiselect(
@@ -254,18 +310,15 @@ if __name__ == "__main__":
             )
 
             st.session_state.limit_param = None
-            # check box for limit
-            sidebar.checkbox(
-                label="Limit",
-                key="limit_yes",
-            )
-            if st.session_state.limit_yes:
-                # show radio to choose between specific limit or all
-                limit_type = sidebar.radio(
-                    label="Limit Type", options=["Specific", "All"], horizontal=True, key="limit_type", label_visibility="collapsed"
+
+            if st.session_state.method_params.limit:
+                # check box for limit
+                sidebar.toggle(
+                    label="Download All",
+                    value=False,
+                    key="download_all",
                 )
-                if limit_type == "Specific":
-                    # show the limit for first limit parameter as defined in the method file
+                if not st.session_state.download_all:
                     limit = sidebar.number_input(
                         label=f"**{st.session_state.method_params.limit[0]} limit:**",
                         value=100,
@@ -307,25 +360,30 @@ if __name__ == "__main__":
 
             for param in st.session_state.parameters:
                 st.subheader(f"**{param}**:")
-                st.write(st.session_state.method_params.parameters[param]["description"])
+                param_info = xbrl.define(param)
+                st.markdown(param_info["description"])
+                type = param_info["type"]
 
-                if st.session_state.method_params.parameters[param]["type"] == "boolean":
+                if type == "boolean":
                     boolean_input_for_booleans(param)
 
-                elif st.session_state.method_params.parameters[param]["type"] == "integer":
+                elif type == "integer":
                     input_number_for_integers(param)
 
-                elif st.session_state.method_params.parameters[param]["type"] == "string":
+                elif type == "string":
                     text_input_for_strings(param)
 
-                elif st.session_state.method_params.parameters[param]["type"] == "array[integer]":
+                elif type == "array[integer]":
                     range_and_slider_for_array_integers(
                         param,
-                        st.session_state.method_params.parameters[param]["values"],
+                        param_info["values"],
                     )
 
-                elif st.session_state.method_params.parameters[param]["type"] == "array[string]":
-                    text_box_for_array_strings_no_ops(param)
+                elif type == "array[string]":
+                    text_box_for_array_strings_no_ops(
+                        param,
+                        param_info["placeholder"],
+                    )
 
             st.session_state.query_params = {"fields": st.session_state.fields}
 
@@ -342,109 +400,105 @@ if __name__ == "__main__":
                 st.session_state.query_params["unique"] = True
             st.session_state.query_params["method"] = method
 
-    # create a checkbox to show the query parameters
-    st.checkbox(
-        label="Show Query Parameters",
-        key="show_query_params",
-        help="Show the query parameters.",
-    )
-    if st.session_state.show_query_params:
-        st.write(st.session_state.query_params)
+        # create a checkbox to show the query parameters
+        st.checkbox(
+            label="Show Query Parameters",
+            key="show_query_params",
+            help="Show the query parameters.",
+        )
+        if st.session_state.show_query_params:
+            st.write(st.session_state.query_params)
 
-    # run the query
-    query_btn_disabled = True
-    if len(st.session_state["fields"]) > 0:
-        query_btn_disabled = False
+        # run the query
+        query_btn_disabled = True
+        if len(st.session_state["fields"]) > 0:
+            query_btn_disabled = False
 
-    query_button_placeholder.button(
-        label="Run Query",
-        key="run_query",
-        type="primary",
-        use_container_width=True,
-        disabled=query_btn_disabled,
-    )
-    new_results_placeholder = st.empty()
-    if st.session_state.run_query:
-        try:
-            with st.spinner("Running query..."):
-                st.session_state.pop("last_query", None)
-                st.session_state.last_query = xbrl.query(
-                    **st.session_state.query_params, as_dataframe=True, print_query=True, streamlit=True
-                )
+        query_button_placeholder.button(
+            label="Run Query",
+            key="run_query",
+            type="primary",
+            use_container_width=True,
+            disabled=query_btn_disabled,
+        )
+        new_results_placeholder = st.empty()
+        if st.session_state.run_query:
+            try:
+                with st.spinner("Running query..."):
+                    st.session_state.pop("last_query", None)
+                    st.session_state.last_query = xbrl.query(
+                        **st.session_state.query_params, as_dataframe=True, print_query=True, streamlit=True, timeout=10
+                    )
+                    if "account_limit" not in st.session_state:
+                        st.session_state.account_limit = xbrl.account_limit
 
-        except Exception as e:
-            new_results_placeholder.warning(
-                f"Your query is taking a long time... \n"
-                f"The server may still be working on this query. "
-                f"You can wait and try again in a few minutes. "
-                f"Or try narrowing your search criteria. \n"
-                f"\n {e}"
-            )
-            st.stop()
+            except Exception as e:
+                new_results_placeholder.error(f"{e}")
+                st.stop()
 
-    with new_results_placeholder.container():
-        # show the dataframe
-        st.subheader("Last Query Results")
-        if "last_query" not in st.session_state:
-            st.info("No **Query** has been run yet.")
+        with new_results_placeholder.container():
+            # show the dataframe
+            st.subheader("Last Query Results")
+            if "last_query" not in st.session_state:
+                st.info("No **Query** has been submitted yet")
 
-        else:
-            # show a download button to get the data in csv format
-            # box for file name
-            filename = st.text_input(
-                label="File Name",
-                value="xbrl data",
-            )
-            dwnld_btn_place, del_btn_place = st.columns(2)
-
-            # show a button to show the full data
-            st.checkbox(
-                label="My computer rocks! 🚀 Show Full Data",
-                help="Show the full data.",
-                key="show_full_data",
-            )
-            if st.session_state.show_full_data:
-                st.success(
-                    f"""Viewing full data: **{st.session_state.last_query.shape[0]}**
-                    rows and **{st.session_state.last_query.shape[1]}** columns."""
-                )
-
-                st.dataframe(
-                    data=st.session_state.last_query,
-                    use_container_width=True,
-                    hide_index=True,
-                )
             else:
-                st.success(
-                    f"""Query has **{st.session_state.last_query.shape[0]}** rows.
-                    You are viewing **{min(100, st.session_state.last_query.shape[0])}** rows
-                    and **{st.session_state.last_query.shape[1]}** columns.
-                    You can try **Show Full Data** or **Download** the full data instead."""
+                # show a download button to get the data in csv format
+                # box for file name
+                filename = st.text_input(
+                    label="File Name",
+                    value="xbrl data",
                 )
+                dwnld_btn_place, del_btn_place = st.columns(2)
 
-                st.dataframe(
-                    data=st.session_state.last_query.head(100),
-                    use_container_width=True,
-                    hide_index=True,
+                # show a button to show the full data
+                st.checkbox(
+                    label="My computer rocks! 🚀 Show Full Data",
+                    help="Show the full data.",
+                    key="show_full_data",
                 )
+                if st.session_state.show_full_data:
+                    st.success(
+                        f"""Viewing full data: **{st.session_state.last_query.shape[0]}**
+                        rows and **{st.session_state.last_query.shape[1]}** columns."""
+                    )
 
-            with dwnld_btn_place:
-                st.download_button(
-                    label="Download as CSV File",
-                    use_container_width=True,
-                    data=st.session_state.last_query.to_csv(index=False).encode("utf-8"),
-                    file_name=f"{filename}.csv",
-                    mime="text/csv",
-                    key="download_data",
-                )
+                    st.dataframe(
+                        data=st.session_state.last_query,
+                        use_container_width=True,
+                        hide_index=True,
+                    )
+                else:
+                    st.success(
+                        f"""Query has **{st.session_state.last_query.shape[0]}** rows.
+                        You are viewing **{min(100, st.session_state.last_query.shape[0])}** rows
+                        and **{st.session_state.last_query.shape[1]}** columns.
+                        You can try **Show Full Data** or **Download** the full data instead."""
+                    )
 
-            with del_btn_place:
-                st.button(
-                    label="Delete Query",
-                    key="delete_query_btn",
-                    on_click=lambda: st.session_state.pop("last_query"),
-                    type="primary",
-                    use_container_width=True,
-                )
+                    st.dataframe(
+                        data=st.session_state.last_query.head(100),
+                        use_container_width=True,
+                        hide_index=True,
+                    )
 
-    # st.write(st.session_state)
+                with dwnld_btn_place:
+                    st.download_button(
+                        label="Download as CSV File",
+                        use_container_width=True,
+                        data=st.session_state.last_query.to_csv(index=False).encode("utf-8"),
+                        file_name=f"{filename}.csv",
+                        mime="text/csv",
+                        key="download_data",
+                    )
+
+                with del_btn_place:
+                    st.button(
+                        label="Delete Query",
+                        key="delete_query_btn",
+                        on_click=lambda: st.session_state.pop("last_query"),
+                        type="primary",
+                        use_container_width=True,
+                    )
+
+        # st.write(st.session_state)